PlayerPRO Player: have the error value returned if the old importer returns a bad value be a Cocoa error value. This makes it so we don't have to include PlayerPROKit on that source file. Documentation changes/additions.

+++ PlayerPRO 6/Plug-ins/Advanced Import plug-ins/Classic Application/ImportWindowController.swift
@@ -65,7 +65,7 @@ class ImportWindowController: NSWindowController {
 				
 				var unusedDriverSettings = MADDriverSettings.new()
 				madMusic = UnsafeMutablePointer<MADMusic>.alloc(1)
-				errVal = madLoad(UnsafePointer<Int8>(aData.bytes), size_t(aData.length), madMusic, &unusedDriverSettings)
+				errVal = madLoad(UnsafePointer<Int8>(aData.bytes), aData.length, madMusic, &unusedDriverSettings)
 				
 				if errVal != .NoErr {
 					// The importers should have cleaned up after themselves...
